+++ src/lib/superStruct.ts
@@ -1,13 +1,15 @@
 export type SuperStructArray<T extends Record<string, unknown>, Data, ArrayData = unknown> = {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	[Property in keyof T]?: T extends any
-		? T[Property] extends Record<string, unknown>
-			? SuperStructArray<T[Property], Data, ArrayData>
-			: T[Property] extends (infer A)[]
+		? NonNullable<T[Property]> extends Record<string, unknown>
+			? SuperStructArray<NonNullable<T[Property]>, Data, ArrayData>
+			: NonNullable<T[Property]> extends (infer A)[]
 				? ArrayData &
 						Record<
 							number,
-							A extends Record<string, unknown> ? SuperStructArray<A, Data, ArrayData> : Data
+							NonNullable<A> extends Record<string, unknown>
+								? SuperStructArray<NonNullable<A>, Data, ArrayData>
+								: Data
 						>
 				: Data
 		: never;
@@ -16,11 +18,11 @@ export type SuperStructArray<T extends Record<string, unknown>, Data, ArrayData 
 export type SuperStruct<T extends Record<string, unknown>, Data> = Partial<{
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	[Property in keyof T]: T extends any
-		? T[Property] extends Record<string, unknown>
-			? SuperStruct<T[Property], Data>
-			: T[Property] extends (infer A)[]
-				? A extends Record<string, unknown>
-					? SuperStruct<A, Data>
+		? NonNullable<T[Property]> extends Record<string, unknown>
+			? SuperStruct<NonNullable<T[Property]>, Data>
+			: NonNullable<T[Property]> extends (infer A)[]
+				? NonNullable<A> extends Record<string, unknown>
+					? SuperStruct<NonNullable<A>, Data>
 					: Data
 				: Data
 		: never;